Fixed warnings on Linux: - removed warning messages on g++ (-pedantic -Wall) - added missing return statements in non-void methods - fixed signed/unsigned mismatches

+++ src/geSG/AttribConfig.cpp
@@ -55,8 +55,8 @@ bool AttribConfig::allocData(AttribReference &r,int numVertices,int numIndices,i
    AttribStorageList::iterator storageIt=_attribStorages.end();
    for(auto it=_attribStorages.begin(); it!=_attribStorages.end(); it++)
    {
-      if((*it)->getNumVerticesAvailableAtTheEnd()>=numVertices &&
-         (*it)->getNumIndicesAvailableAtTheEnd()>=numIndices)
+      if((*it)->getNumVerticesAvailableAtTheEnd()>=unsigned(numVertices) &&
+         (*it)->getNumIndicesAvailableAtTheEnd()>=unsigned(numIndices))
       {
          storageIt=it;
          break;
